requirements.txt
crewai[anthropic]
crewai-tools
langchain-community
ddgs
python-dotenv
anthropic
python-telegram-bot[ext]>=20
